common: drop redundant intermediate casts surfaced by SignExtend/Truncate fold
After the SignExtend/ZeroExtend/Truncate helpers were inlined as explicit
static_cast chains, a handful of sites ended up with a redundant
`static_cast<uN>` intermediate between the source and a `static_cast<sN>`
(or s64) outer cast. The pattern is:

    static_cast<sN>(static_cast<uM>(EXPR))

where EXPR's natural type already fits in sN. In those cases the uM step
is pure noise - it widens to an unsigned type that's then immediately
reinterpreted as signed of equal or greater width, while EXPR was already
a sub-width unsigned type that converts to sN value-preservingly through
either integer promotion or a direct cast.

Sites collapsed:

  gpu_sw_backend.h:61-63 (SetRGB24Dithered)
    `static_cast<s32>(static_cast<u32>(rN)) + offset` -> `static_cast<s32>(rN) + offset`
    where rN is u8. u8 widens to s32 value-preservingly.

  mdec.cpp:583, 604 (DecodeRLE_Old)
    `static_cast<s32>(static_cast<u32>(qt[i]))` -> `static_cast<s32>(qt[i])`
    qt is u8*; same reasoning.

  cdrom.h:234 (ApplyVolume)
    `static_cast<s32>(static_cast<u32>(volume))` -> `static_cast<s32>(volume)`
    volume is u8.

  gte.cpp:452 (UNRDivide)
    `static_cast<s32>(static_cast<u32>(divisor))` -> `static_cast<s32>(divisor)`
    divisor is already u32; the inner u32 cast is a no-op.

  gte.cpp:663 (NCDS)
    `static_cast<s64>(static_cast<u64>(UNRDivide(...)))` -> `static_cast<s64>(UNRDivide(...))`
    UNRDivide returns u32; u32 widens to s64 value-preservingly.

  cpu_recompiler_code_generator_aarch64.cpp:1087, 1115, 1142
    (EmitLoadCPUStructField, EmitStoreCPUStructField, EmitLoadGuestRegister)
    `static_cast<s64>(static_cast<u64>(offset))` -> `static_cast<s64>(offset)`
    offset is u32 parameter; same reasoning. Three identical sites.

Codegen verification: built a small test harness that exercises each
collapsed pattern alongside its original form. -O2 g++ x86_64 emits
byte-identical assembly for every pair:

  u8 widen + s32 add:    movzbl ; leal           (both)
  self-cast u32:         movl                    (both)
  u32 -> s64 widen:      movl                    (both, mov implicitly zero-extends on x86-64)
  u8 -> s32 * s16 -> sar: movzbl ; movswl ; imull ; movl ; sarl   (both)

These are pure source-readability cleanups - no semantic change, no
warning delta. Spotted while scanning for follow-up opportunities after
the helper-removal commit landed; the patterns become more visible once
the inner step is no longer hidden behind a `ZeroExtend32` function call.

+++ src/core/cdrom.h
@@ -231,7 +231,7 @@ private:
   }
   ALWAYS_INLINE static constexpr s32 ApplyVolume(s16 sample, u8 volume)
   {
-    return s32(sample) * static_cast<s32>(static_cast<u32>(volume)) >> 7;
+    return s32(sample) * static_cast<s32>(volume) >> 7;
   }
 
   ALWAYS_INLINE static constexpr s16 SaturateVolume(s32 volume)
